+++ examples/cfd/lagrangian_mgn/inference.py
@@ -14,56 +14,53 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import logging
 import os
-import time
+
 import hydra
-from hydra.utils import to_absolute_path
+from hydra.utils import instantiate, to_absolute_path
 
 import dgl
 from dgl.dataloading import GraphDataLoader
-import matplotlib.pyplot as plt
+
+import matplotlib
 from matplotlib import animation
-from matplotlib import tri as mtri
-from matplotlib.patches import Rectangle
-import matplotlib  #
+from matplotlib import pyplot as plt
 
 matplotlib.use("TkAgg")  # for plotting
 
-import numpy as np
-from networkx import radius
-from omegaconf import DictConfig
+from omegaconf import DictConfig, OmegaConf
+
 import torch
 
-from modulus.models.meshgraphnet import MeshGraphNet
-from modulus.datapipes.gnn.lagrangian_dataset import LagrangianDataset, graph_update
-from modulus.launch.logging import PythonLogger
+from modulus.datapipes.gnn.lagrangian_dataset import graph_update
 from modulus.launch.utils import load_checkpoint
+
+from loggers import get_gpu_info, init_python_logging
+
+
+logger = logging.getLogger("lmgn")
 
 
 class MGNRollout:
-    def __init__(self, cfg: DictConfig, logger: PythonLogger):
-        self.num_test_samples = cfg.num_test_samples
-        self.num_test_time_steps = cfg.num_test_time_steps
-        self.dim = cfg.num_output_features
-        self.frame_skip = cfg.frame_skip
+    def __init__(self, cfg: DictConfig):
+        self.num_steps = cfg.data.test.num_steps
+        self.dim = cfg.dim
+        self.frame_skip = cfg.inference.frame_skip
         self.num_history = 5
-        self.num_node_type = 6
+        self.num_node_type = cfg.data.test.num_node_types
         self.plotting_index = 0
-        self.radius = cfg.radius
+        self.radius = cfg.data.test.radius
 
         # set device
-        self.device = "cuda" if torch.cuda.is_available() else "cpu"
+        self.device = cfg.test.device
         logger.info(f"Using {self.device} device")
 
         # instantiate dataset
-        self.dataset = LagrangianDataset(
-            name="Water",
-            data_dir=to_absolute_path(cfg.data_dir),
-            split="valid",
-            num_samples=cfg.num_test_samples,
-            num_steps=cfg.num_test_time_steps,
-            radius=cfg.radius,
-        )
+        logger.info("Loading the test dataset...")
+        self.dataset = instantiate(cfg.data.test)
+        logger.info(f"Using {len(self.dataset)} test samples.")
+
         self.dim = self.dataset.dim
         self.dt = self.dataset.dt
         self.bound = self.dataset.bound
@@ -80,33 +77,24 @@ class MGNRollout:
         # instantiate dataloader
         self.dataloader = GraphDataLoader(
             self.dataset,
-            batch_size=1,
-            shuffle=False,
-            drop_last=False,
+            **cfg.test.dataloader,
         )
 
         # instantiate the model
-        self.model = MeshGraphNet(
-            cfg.num_input_features,
-            cfg.num_edge_features,
-            cfg.num_output_features,
-            cfg.processor_size,
-            mlp_activation_fn=cfg.activation,
-            do_concat_trick=cfg.do_concat_trick,
-            num_processor_checkpoint_segments=cfg.num_processor_checkpoint_segments,
-            recompute_activation=cfg.recompute_activation,
-        )
-        if cfg.jit:
-            self.model = torch.jit.script(self.model).to(self.device)
-        else:
-            self.model = self.model.to(self.device)
+        logger.info("Creating the model...")
+        # instantiate the model
+        self.model = instantiate(cfg.model)
+
+        if cfg.compile.enabled:
+            self.model = torch.compile(self.model, **cfg.compile.args)
+        self.model = self.model.to(self.device)
 
         # enable train mode
         self.model.eval()
 
         # load checkpoint
         load_checkpoint(
-            to_absolute_path(cfg.ckpt_path),
+            to_absolute_path(cfg.resume_dir),
             models=self.model,
             device=self.device,
         )
@@ -248,7 +236,7 @@ class MGNRollout:
 
     def animate2d(self, num):
         num *= self.frame_skip
-        num = num + self.plotting_index * self.num_test_time_steps
+        num = num + self.plotting_index * self.num_steps
         node_type = self.node_type[num]
         node_type = (
             torch.argmax(node_type, dim=1).numpy() / self.num_node_type
@@ -290,7 +278,7 @@ class MGNRollout:
 
     def animate3d(self, num):
         num *= self.frame_skip
-        num = num + self.plotting_index * self.num_test_time_steps
+        num = num + self.plotting_index * self.num_steps
         node_type = self.node_type[num]
         node_type = (
             torch.argmax(node_type, dim=1).numpy() / self.num_node_type
@@ -335,28 +323,32 @@ class MGNRollout:
         return plt
 
 
-@hydra.main(version_base="1.3", config_path="conf", config_name="config_2d")
+@hydra.main(version_base="1.3", config_path="conf", config_name="config")
 def main(cfg: DictConfig) -> None:
-    logger = PythonLogger("main")  # General python logger
-    logger.file_logging()
+    init_python_logging(cfg, base_filename="inference")
+    logger.info(f"Config summary:\n{OmegaConf.to_yaml(cfg, sort_keys=True)}")
+    logger.info(get_gpu_info())
+
     logger.info("Rollout started...")
-    rollout = MGNRollout(cfg, logger)
+    rollout = MGNRollout(cfg)
 
     # test on dataset
     rollout.predict()
 
     # unit test
-    # rollout.unit_test_example(t=cfg.num_test_time_steps)
+    # rollout.unit_test_example(t=cfg.num_steps)
 
     # compute the roll out loss
     pred = torch.stack([tensor.reshape(-1) for tensor in rollout.pred], dim=0)
     target = torch.stack([tensor.reshape(-1) for tensor in rollout.exact], dim=0)
     loss = torch.nn.functional.mse_loss(pred, target)
-    print(f"the rollout loss is {loss}")
+    logger.info(f"The rollout loss is {loss:.5f}")
 
     # plot the roll out loss
     error_plt = rollout.plot_error(pred, target)
-    error_plt.savefig("animations/error.png")
+    out_dir = os.path.join(cfg.output, "animations")
+    os.makedirs(out_dir, exist_ok=True)
+    error_plt.savefig(os.path.join(out_dir, "error.png"))
 
     # plot
     if cfg.dim == 2:
@@ -364,19 +356,19 @@ def main(cfg: DictConfig) -> None:
         ani = animation.FuncAnimation(
             rollout.fig,
             rollout.animate2d,
-            frames=(cfg.num_test_time_steps - 5) // cfg.frame_skip,
-            interval=cfg.frame_interval,
+            frames=(cfg.data.test.num_steps - 5) // cfg.inference.frame_skip,
+            interval=cfg.inference.frame_interval,
         )
     elif cfg.dim == 3:
         rollout.init_animation3d(index=0)
         ani = animation.FuncAnimation(
             rollout.fig,
             rollout.animate3d,
-            frames=(cfg.num_test_time_steps - 5) // cfg.frame_skip,
-            interval=cfg.frame_interval,
+            frames=(cfg.data.test.num_steps - 5) // cfg.inference.frame_skip,
+            interval=cfg.inference.frame_interval,
         )
 
-    ani.save("animations/animation.gif")
+    ani.save(os.path.join(out_dir, "animation.gif"))
     logger.info(f"Created animation")